Move ExpressionUser to its own file
