+++ fcd/ast/expression_use.h
@@ -114,158 +114,4 @@ public:
 	operator const Expression*() const { return getUse(); }
 };
 
-class ExpressionUser
-{
-	template<bool B, typename T>
-	using OptionallyConst = typename std::conditional<B, typename std::add_const<T>::type, typename std::remove_const<T>::type>::type;
-	
-public:
-	enum UserType : unsigned
-	{
-		// statements
-		StatementMin,
-		Noop = StatementMin,
-		Sequence,
-		IfElse,
-		Loop,
-		Expr,
-		Keyword,
-		StatementMax,
-		
-		// expressions
-		ExpressionMin = StatementMax,
-		Token = ExpressionMin,
-		UnaryOperator,
-		NAryOperator,
-		Call,
-		Cast,
-		Numeric,
-		Ternary,
-		Aggregate,
-		Subscript,
-		Assembly,
-		Assignable,
-		ExpressionMax,
-	};
-	
-	template<bool IsConst>
-	class UseIterator : public std::iterator<std::forward_iterator_tag, OptionallyConst<IsConst, ExpressionUse>>
-	{
-		const ExpressionUseAllocInfo* allocInfo;
-		OptionallyConst<IsConst, ExpressionUse>* useListEnd;
-		unsigned index;
-		
-		void goToNextNonEmptyBuffer();
-		
-	public:
-		UseIterator(std::nullptr_t)
-		: allocInfo(nullptr), useListEnd(nullptr), index(0)
-		{
-		}
-		
-		UseIterator(OptionallyConst<IsConst, ExpressionUser>* user)
-		: allocInfo(&user->allocInfo), index(0)
-		{
-			useListEnd = reinterpret_cast<OptionallyConst<IsConst, ExpressionUse>*>(user);
-			goToNextNonEmptyBuffer();
-		}
-		
-		UseIterator(const UseIterator&) = default;
-		UseIterator(UseIterator&&) = default;
-		
-		OptionallyConst<IsConst, ExpressionUse>& operator*() { return *operator->(); }
-		OptionallyConst<IsConst, ExpressionUse>* operator->() { return useListEnd - index - 1; }
-		
-		template<bool B>
-		bool operator==(const UseIterator<B>& that) const { return useListEnd == that.useListEnd && index == that.index; }
-		
-		template<bool B>
-		bool operator!=(const UseIterator<B>& that) const { return !(*this == that); }
-		
-		UseIterator& operator++();
-		UseIterator operator++(int)
-		{
-			UseIterator copy = *this;
-			operator++();
-			return copy;
-		}
-	};
-	
-	typedef UseIterator<false> iterator;
-	typedef UseIterator<true> const_iterator;
-	
-private:
-	ExpressionUseAllocInfo allocInfo;
-	UserType userType;
-	
-	// force class to have a vtable
-	virtual void anchor();
-	
-protected:
-	ExpressionUse& insertUseAtEnd();
-	iterator erase(iterator iter); // protected because not every class will want to expose it
-	
-public:
-	ExpressionUser(UserType type, unsigned allocatedUses, unsigned usedUses)
-	: userType(type), allocInfo(allocatedUses, usedUses)
-	{
-	}
-	
-	ExpressionUser(UserType type, unsigned inlineUses)
-	: ExpressionUser(type, inlineUses, inlineUses)
-	{
-	}
-	
-	UserType getUserType() const { return userType; }
-	
-	ExpressionUse& getOperandUse(unsigned index);
-	const ExpressionUse& getOperandUse(unsigned index) const { return const_cast<ExpressionUser*>(this)->getOperandUse(index); }
-	
-	Expression* getOperand(unsigned index) { return getOperandUse(index).getUse(); }
-	const Expression* getOperand(unsigned index) const { return getOperandUse(index).getUse(); }
-	
-	void setOperand(unsigned index, Expression* expression) { getOperandUse(index).setUse(expression); }
-	
-	unsigned operands_size() const;
-	const_iterator operands_begin() const { return const_iterator(this); }
-	const_iterator operands_cbegin() const { return operands_begin(); }
-	const_iterator operands_end() const { return const_iterator(nullptr); }
-	const_iterator operands_cend() const { return operands_end(); }
-	iterator operands_begin() { return iterator(this); }
-	iterator operands_end() { return iterator(nullptr); }
-	llvm::iterator_range<iterator> operands() { return llvm::make_range(operands_begin(), operands_end()); }
-	llvm::iterator_range<const_iterator> operands() const { return llvm::make_range(operands_begin(), operands_end()); }
-	
-	void print(llvm::raw_ostream& os) const;
-	void dump() const;
-};
-
-template<bool IsConst>
-void ExpressionUser::UseIterator<IsConst>::goToNextNonEmptyBuffer()
-{
-	while (useListEnd != nullptr && index == allocInfo->used)
-	{
-		auto useListBegin = useListEnd - allocInfo->allocated;
-		auto& arrayHead = reinterpret_cast<OptionallyConst<IsConst, ExpressionUseArrayHead>*>(useListBegin)[-1];
-		useListEnd = &arrayHead.array[arrayHead.allocInfo.allocated];
-		allocInfo = &arrayHead.allocInfo;
-		index = 0;
-	}
-}
-
-template<bool IsConst>
-ExpressionUser::UseIterator<IsConst>& ExpressionUser::UseIterator<IsConst>::operator++()
-{
-	index++;
-	goToNextNonEmptyBuffer();
-	return *this;
-}
-
-#define OPERAND_GET_SET_T(name, type, index) \
-	NOT_NULL(type) get##name() { return llvm::cast<type>(getOperand(index)); } \
-	NOT_NULL(const type) get##name() const { return llvm::cast<type>(getOperand(index)); } \
-	void set##name(NOT_NULL(type) op) { getOperandUse(index).setUse(op); }
-
-#define OPERAND_GET_SET(name, index) OPERAND_GET_SET_T(name, Expression, index)
-
 #endif /* use_list_hpp */
